requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-phonenumber-field==7.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
phonenumberslite==8.12.55
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
